+++ src/lib/accessibility/FocusZone/CHANGELOG.md
@@ -4,8 +4,7 @@ This is a list of changes made to this Stardust copy of FocusZone in comparison 
 
 ### feat(FocusZone): Implement FocusZone into renderComponent [#116](https://github.com/stardust-ui/react/pull/116)
 - Prettier and linting fixes, e.g., removing semicolons, removing underscores from private methods.
-- Replaced path `../../Utilities` with `@uifabric/utilities`.
-- Moved `IS_FOCUSABLE_ATTRIBUTE` and others to `FocusUtilities.ts`.
+- Moved `IS_FOCUSABLE_ATTRIBUTE` and others to `focusUtilities.ts`.
 - Added prop types.
 - Added `preventDefaultWhenHandled` property and method, added to `_onKeyDown`.
 - Renamed boolean callback properties to better reflect their purpose.
@@ -26,4 +25,4 @@ This is a list of changes made to this Stardust copy of FocusZone in comparison 
     - `getParent` replaced with vanilla JS approach.
     - `getRTL` replaced with an `isRtl` property.
     - `createRef` replaced with a custom object and a callback which is necessary anyway because of custom component handling, see above for details.
-    - Focus related utilities moved to `FocusUtilities.ts`.
+    - Focus related utilities moved to `focusUtilities.ts`.